Refactor the whole Project into ASM, IR and AR
The whole project is now refactored into 3 distinct representations of the
data. The Assembly is a raw representation of the input which is then
being converted into an Intermediate Representation, which gets optimized
and later converted into Action Replay Codes. By introducing an
intermediate representation, WiiRD codes or Rom Hacks would be possible
too in the future, without having to change too much.