Introduce specialized Optimizers for the Writers
The Writers might need to optimize the Intermediate Representation before
writing out the individual instructions. Action Replay for example
represents the Bit Setting Instructions as normal writes. Those can
potentially be merged into other writes though, which is one of the
reasons Writers now get their own Optimization Stage. After they optimize
the Intermediate Representation, the General Optimizer is being run again
to ensure that the Intermediate Representation is as optimal, as it can
be.